requirements.txt
playwright>=1.45.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0 